pyproject.toml
[project]
name = "pgvector-lab"
version = "0.1.0"
description = "A lab environment for experimenting with pgvector"
authors = [{ name = "Emiiano Frigo" }]
dependencies = [
    "numpy==2.0.2",
    "langchain-ollama==0.3.2",
    "docling>=2.31.0",
    "pyspark==3.5.1",
    "langchain==0.3.25",
    "ipykernel>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "pgvector>=0.4.1",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["black==24.3.0", "isort==5.10.1", "pylint==3.0.3", "pre-commit==3.6.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
